// src/app/core/models/estate-units/estate-unit.ts
export class EstateUnits {
    id?: number;
    estateUnitName: string;
    cityId: number;
    stateId: number;
    countryId: number;
    address: string;
    nit: string;
    cadastreNumber: string;
    identificationTypeId: number;
    identificationNumber: number;
    name: string;
    lastName: string;
    phone: string;
    mobile: string;
    email: string;
    numberGroup: number;
    numberIndependentUnit: number;
    numberPedestrianEntrance: number;
    numberVehicularEntrance: number;
    numberParkingLots: number;
    status: number;
    idRepresentative: number;
}